x/bridge/client/cli/query_register_keygen.go
package cli

import (
	"context"

	"bridge/x/bridge/types"
	"github.com/cosmos/cosmos-sdk/client"
	"github.com/cosmos/cosmos-sdk/client/flags"
	"github.com/spf13/cobra"
)

func CmdListRegisterKeygen() *cobra.Command {
	cmd := &cobra.Command{
		Use:   "list-register-keygen",
		Short: "list all RegisterKeygen",
		RunE: func(cmd *cobra.Command, args []string) error {
			clientCtx := client.GetClientContextFromCmd(cmd)

			pageReq, err := client.ReadPageRequest(cmd.Flags())
			if err != nil {
				return err
			}

			queryClient := types.NewQueryClient(clientCtx)

			params := &types.QueryAllRegisterKeygenRequest{
				Pagination: pageReq,
			}

			res, err := queryClient.RegisterKeygenAll(context.Background(), params)
			if err != nil {
				return err
			}

			return clientCtx.PrintProto(res)
		},
	}

	flags.AddPaginationFlagsToCmd(cmd, cmd.Use)
	flags.AddQueryFlagsToCmd(cmd)

	return cmd
}

func CmdShowRegisterKeygen() *cobra.Command {
	cmd := &cobra.Command{
		Use:   "show-register-keygen [index]",
		Short: "shows a RegisterKeygen",
		Args:  cobra.ExactArgs(1),
		RunE: func(cmd *cobra.Command, args []string) (err error) {
			clientCtx := client.GetClientContextFromCmd(cmd)

			queryClient := types.NewQueryClient(clientCtx)

			argIndex := args[0]

			params := &types.QueryGetRegisterKeygenRequest{
				Index: argIndex,
			}

			res, err := queryClient.RegisterKeygen(context.Background(), params)
			if err != nil {
				return err
			}

			return clientCtx.PrintProto(res)
		},
	}

	flags.AddQueryFlagsToCmd(cmd)

	return cmd
}
